OSIS-6247 : Move to learning_unit and rename files/tests
